refactor(channel): move runtime_state.rs to runtime/state.rs
Move channel runtime state module from channel/runtime_state.rs to
channel/runtime/state.rs and update all affected imports across the
channel submodules (dispatch, registry, types, commands, wecom, feishu,
whatsapp, etc.).

Also updated runtime/serve.rs to import from the new location.
