+++ crates/app/src/channel/runtime/mod.rs
@@ -2,4 +2,5 @@
 
 pub(in crate::channel) mod http;
 pub(in crate::channel) mod serve;
+pub(in crate::channel) mod state;
 pub(in crate::channel) mod turn_feedback;
