+++ crates/app/src/channel/types.rs
@@ -52,7 +52,7 @@ use serde_json::Value;
     feature = "channel-wecom",
     feature = "channel-whatsapp"
 ))]
-use super::runtime_state::ChannelOperationRuntimeTracker;
+use super::runtime::state::ChannelOperationRuntimeTracker;
 #[cfg(any(
     feature = "channel-telegram",
     feature = "channel-feishu",
