+++ crates/app/src/channel/whatsapp/mod.rs
@@ -14,7 +14,7 @@ use super::runtime::serve::{ChannelServeCommandSpec, ChannelServeStopHandle};
 use super::{ChannelOutboundTargetKind, WHATSAPP_COMMAND_FAMILY_DESCRIPTOR};
 use super::{
     http::{ChannelOutboundHttpPolicy, build_outbound_http_client, validate_outbound_http_target},
-    runtime_state::ChannelOperationRuntimeTracker,
+    runtime::state::ChannelOperationRuntimeTracker,
 };
 use crate::config::{ChannelDefaultAccountSelectionSource, LoongClawConfig};
 use crate::{CliResult, KernelContext, config::ResolvedWhatsappChannelConfig};
